Correct JSON decoding and Add Expired Option
The JSON fix shouldn't be necessary any more after recent changes by crt.sh.
https://github.com/crtsh/certwatch_db/commit/f4f46ea37c23543c4cdf1a3c8867d68967641807

Added an option to exclude expired certificates. The default is to include expired certificates to keep current functionality.

For wildcard checks, changed the %25 to %, and added a check for wildcards already in the domain query string.